Workaround for number pad keyboard on iPhone.
Add a simple tap gesture recognizer on the table view that will end
editing if tapping on the table view. This will hide the number pad on
the iPhone which does not have a way to dismiss the keyboard as soon as
you tap on another field that has no keyboard (e.g. a slider or switch)
or into the background.