fix occasional crash when pushing child-viewcontroller
from time to time, when the NavigationController can't keep up,
IASK crashes with UIKit complaining about pushing multiple
instances of the same view controller. Fix that by removing
the view controller caching.
To still be able to push the hidden keys down the stack, keep
a (weak) reference to the currently presented childViewController.